Move TransactionDb to Core
