+++ wallet/src/main/scala/org/bitcoins/wallet/models/IncomingTransactionDAO.scala
@@ -1,6 +1,7 @@
 package org.bitcoins.wallet.models
 
 import org.bitcoins.core.currency.CurrencyUnit
+import org.bitcoins.core.wallet.transactions.TransactionDb
 import org.bitcoins.crypto.DoubleSha256DigestBE
 import org.bitcoins.wallet.config._
 import slick.lifted.{PrimaryKey, ProvenShape}
