+++ wallet/src/main/scala/org/bitcoins/wallet/models/OutgoingTransactionDAO.scala
@@ -2,6 +2,7 @@ package org.bitcoins.wallet.models
 
 import org.bitcoins.core.currency.CurrencyUnit
 import org.bitcoins.core.wallet.fee.SatoshisPerByte
+import org.bitcoins.core.wallet.transactions.TransactionDb
 import org.bitcoins.crypto.DoubleSha256DigestBE
 import org.bitcoins.wallet.config._
 import slick.lifted.{PrimaryKey, ProvenShape}
