+++ wallet/src/main/scala/org/bitcoins/wallet/models/OutgoingTransactionDb.scala
@@ -3,6 +3,7 @@ package org.bitcoins.wallet.models
 import org.bitcoins.core.currency._
 import org.bitcoins.core.protocol.transaction._
 import org.bitcoins.core.wallet.fee.SatoshisPerByte
+import org.bitcoins.core.wallet.transactions.TxDB
 import org.bitcoins.crypto.{DoubleSha256Digest, DoubleSha256DigestBE}
 
 /**
